requirements.txt
streamlit
assemblyai
google-generativeai